requirements.txt
sqlalchemy
bcrypt==3.2.0
SQLAlchemy==1.3.13
PyJWT==1.4.2
gunicorn==20.1.0
Flask==1.1.4
psycopg2_binary==2.9.1
flask-sqlalchemy==2.4.1
flask_jwt==0.3.2
Flask-Migrate==3.1.0
requests==2.25.1
flask-restplus==0.13.0
Werkzeug==0.16.0
flask_cors==3.0.10
pycryptodome==3.11.0
sqlalchemy-datatables
markupsafe==2.0.1
# transitions
